requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
boto3==1.35.49
botocore==1.35.49
Django==5.1.1
django-storages==1.14.4
djangorestframework==3.15.2
jmespath==1.0.1
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
statsd==4.0.1
urllib3==2.2.3
